Rename config pkg to azureconfig
