+++ pkg/azure/azureconfig/config.go
@@ -1,10 +1,6 @@
-package config
+package azureconfig
 
 type Config struct {
 	// Env is the environment variable from which the connection string should be loaded.
 	Env string
-}
-
-func New() Config {
-	return Config{}
 }
