+++ pkg/azure/blob/plugin.go
@@ -3,7 +3,7 @@ package blob
 import (
 	"os"
 
-	"get.porter.sh/plugin/azure/pkg/azure/config"
+        "get.porter.sh/plugin/azure/pkg/azure/azureconfig"
 	"get.porter.sh/porter/pkg/instance-storage/claimstore"
 	"github.com/cnabio/cnab-go/utils/crud"
 	"github.com/hashicorp/go-hclog"
@@ -23,7 +23,7 @@ type Plugin struct {
 	CrudStore
 }
 
-func NewPlugin(cfg config.Config) plugin.Plugin {
+func NewPlugin(cfg azureconfig.Config) plugin.Plugin {
 	logger := hclog.New(&hclog.LoggerOptions{
 		Name:   PluginKey,
 		Output: os.Stderr,
